DTSPO-8223 - improve reliability of lock deletion (#40)
* DTSPO-8223 - improve reliability of lock deletion

* DTSPO-8223 - improve reliability of lock deletion

* DTSPO-8223 - testing

* DTSPO-8223 - testing

* DTSPO-8223 - testing

* DTSPO-8223 - testing

* DTSPO-8223 - testing

* DTSPO-8223 - testing

* DTSPO-8223 - add deploy back in

+++ scripts/cleanup-infra.sh
@@ -1,12 +1,28 @@
 #!/usr/bin/env bash
 set -e
 
-# Remove lock on resource group
-echo "Removing lock"
-az group lock delete --subscription "${SUBSCRIPTION}" --resource-group "${SB_RESOURCE_GROUP}" --name "${LOCK_NAME}"
-sleep 5
+get_lock() {
+   az group lock list --subscription "${SUBSCRIPTION}" --resource-group "${SB_RESOURCE_GROUP}" --query '[].name' -o tsv | grep "${LOCK_NAME}"
+}
 
-# Delete PR queue
+if get_lock; then
+  for (( i=0; i<=2; i++ )); do
+    echo "Locks in place: $(get_lock)"
+
+    # Remove lock on resource group
+    echo "Removing lock"
+    az group lock delete --subscription "${SUBSCRIPTION}" --resource-group "${SB_RESOURCE_GROUP}" --name "${LOCK_NAME}"
+
+    if ! get_lock; then
+      break
+    fi
+    sleep 10
+  done
+fi
+
+echo "Lock deleted"
+
+echo "Delete PR queue"
 az servicebus queue delete \
   --namespace-name "${SERVICE_BUS}" \
   --resource-group "${SB_RESOURCE_GROUP}" \
